+++ src/components/CommunityList/Pagination.tsx
@@ -1,26 +1,19 @@
-import {
-  ChevronLeft,
-  ChevronRight,
-  ChevronsLeft,
-  ChevronsRight,
-} from 'lucide-react';
+import { ChevronLeft, ChevronRight, ChevronsLeft, ChevronsRight } from 'lucide-react';
 
 interface PaginationProps {
   page: number;
   totalPages: number;
-  setPage: (page: number) => void; // ✅ 함수형 업데이트 제거
+  onPageChange: (page: number) => void;
 }
 
-export default function Pagination({ page, totalPages, setPage }: PaginationProps) {
+export default function Pagination({ page, totalPages, onPageChange }: PaginationProps) {
   return (
     <div className="flex justify-center items-center gap-2 mt-10">
       {/* 처음으로 */}
       <button
-        className={`p-1 rounded ${
-          page === 1 ? 'text-gray-400 cursor-default' : 'text-black hover:text-purple-600'
-        }`}
+        className={`p-1 rounded ${page === 1 ? 'text-gray-400 cursor-default' : 'text-black hover:text-purple-600'}`}
         disabled={page === 1}
-        onClick={() => setPage(1)}
+        onClick={() => onPageChange(1)}
         aria-label="처음으로"
       >
         <ChevronsLeft size={16} />
@@ -28,11 +21,9 @@ export default function Pagination({ page, totalPages, setPage }: PaginationProp
 
       {/* 이전 페이지 */}
       <button
-        className={`p-1 rounded ${
-          page === 1 ? 'text-gray-400 cursor-default' : 'text-black hover:text-purple-600'
-        }`}
+        className={`p-1 rounded ${page === 1 ? 'text-gray-400 cursor-default' : 'text-black hover:text-purple-600'}`}
         disabled={page === 1}
-        onClick={() => setPage(Math.max(1, page - 1))} // ✅ 함수형 대신 직접 계산
+        onClick={() => onPageChange(Math.max(1, page - 1))}
         aria-label="이전 페이지"
       >
         <ChevronLeft size={16} />
@@ -45,7 +36,7 @@ export default function Pagination({ page, totalPages, setPage }: PaginationProp
         return (
           <button
             key={pageNumber}
-            onClick={() => setPage(pageNumber)}
+            onClick={() => onPageChange(pageNumber)}
             disabled={isCurrent}
             className={`w-8 h-8 text-sm rounded-md transition ${
               isCurrent
@@ -62,13 +53,9 @@ export default function Pagination({ page, totalPages, setPage }: PaginationProp
 
       {/* 다음 페이지 */}
       <button
-        className={`p-1 rounded ${
-          page === totalPages
-            ? 'text-gray-400 cursor-default'
-            : 'text-black hover:text-purple-600'
-        }`}
+        className={`p-1 rounded ${page === totalPages ? 'text-gray-400 cursor-default' : 'text-black hover:text-purple-600'}`}
         disabled={page === totalPages}
-        onClick={() => setPage(Math.min(totalPages, page + 1))} // ✅ 함수형 대신 직접 계산
+        onClick={() => onPageChange(Math.min(totalPages, page + 1))}
         aria-label="다음 페이지"
       >
         <ChevronRight size={16} />
@@ -76,13 +63,9 @@ export default function Pagination({ page, totalPages, setPage }: PaginationProp
 
       {/* 마지막으로 */}
       <button
-        className={`p-1 rounded ${
-          page === totalPages
-            ? 'text-gray-400 cursor-default'
-            : 'text-black hover:text-purple-600'
-        }`}
+        className={`p-1 rounded ${page === totalPages ? 'text-gray-400 cursor-default' : 'text-black hover:text-purple-600'}`}
         disabled={page === totalPages}
-        onClick={() => setPage(totalPages)}
+        onClick={() => onPageChange(totalPages)}
         aria-label="마지막으로"
       >
         <ChevronsRight size={16} />
